Add accuracy detection
It can happend that the location accuracy is very bad. In this case
a presence change can be triggered (i.e. your device thinks that you
are outside the fence due to this fact). Since the accuracy is part
of the owntracks JSON message, it can be used to determine the accuracy
of the event. In the settings page you can set the accuracy in meters.
If the reported accuracy is equal or below the settings value, the triggers
will be fired. Is the accuracy above the settings value, the triggers
will not be fired.