finally solved the multi triggering
When a reconnect event has occured, the on connect is also called.
This causes the on message event to get registered again. So this should
not be called in case of a reconnect.

However i do not think the on message should be inside the on connect event
and should only be registered once, just like the connection.