solved crash in json parsing.
When a message is receivee that does nit contain a message un json
format, the parse function will crash. So solve that in a try/catch
construction so that the exception is caught and further json processing
is cancelled. In the future a new trigger can be added to handle not
json messagea.