Listen to changes in trigger flows.
When a user starts empty, topics are not registered when they are
added in a trigger. This solves that problem. Also if a user changes
the topic, this is also noticed.
Last but not least this way topics that are not used (by changing or
deleting them) are also removed.